refactor: move createOrUpgradeAgentConfiguration to service layer instead of api layer
